requirements.txt
numpy
pandas
PyYAML
opencv-python
albumentations
torch
scipy